Move fakeVolumeHost logic into a separate file
